mcp_learn_agent/requirements.txt
azure-ai-projects
azure-ai-agents>=1.2.0b2
azure-identity
python-dotenv
